Re-arrange test cases.
