+++ component/src/test/resources/testng.xml
@@ -21,11 +21,9 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
 <suite name="Siddhi-Io-Cdc-Test-Suite">
-    <!--Removed the following test:"Siddhi-io-cdc-tests"; since it depends on the local MySQL instance.
-    Uncomment the following test to run with build.-->
     <test name="Siddhi-io-cdc-tests" enabled="true">
         <classes>
-            <class name="org.wso2.extension.siddhi.io.cdc.source.TestCaseOfCDCSource"/>
+            <class name="org.wso2.extension.siddhi.io.cdc.source.TestCaseOfCDCSourceValidation"/>
         </classes>
     </test>
 </suite>
